requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
python-dotenv==1.2.2
pillow==12.1.1
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.48
requests==2.32.5
openai==2.30.0
langchain==1.2.15
pypdf==6.9.2
